backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
frozenlist==1.4.1
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
postgrest==0.16.11
pydantic==2.9.1
pydantic_core==2.23.3
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.2
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
typing_extensions==4.12.2
websockets==12.0
Werkzeug==3.0.4
yarl==1.11.1